Cleaning up after brutal merge of 2.10.x into master.
That's the best I can do. The tests pass, if someone wants
a cleaner merge it is all theirs.

For reference, the merge leading up to this commit was achieved
as follows. The lines with -s ours are where the merge commit
being merged was completely made up of backports from master.

  git merge -s ours eff78b852e  c1dd8bbaa4  && \
  git merge         7026376dcc  ccd7abe897 && \
  git merge -s ours 62681e191a    && \
  git merge         74b3e9aefe 7d80e08469 d24f341f08 c4f49759fe \
    27d73a2352 ba72ee7c6f 42c4cc7a1e d672102fd8 644eb7078a && \
  git merge -s ours 08596af059 b573c287d2   && \
  git merge         d1b6d8b20f    && \
  git merge -s ours 110b54a575 36f78dd606 309ff57ba6  && \
  git merge         06295f9682 d3886086c3 adf51eef76 b403234a27 && \
  git merge -s ours 09d1433064    && \
  git merge         9ddcc1b90e   cabf626bbc && \
  git merge -s ours 283924bfa5