Fix SI-7107: scala now thinks every exception is polymorphic
We need to force info of the `throwableSym` in `addThrowsAnnotation`
because we call `Symbol.isMonomorphicType` that relies on a symbol
being initialized to give right answers.

In the future we should just clean up implementation of
`isMonomorphicType` method to not rely on a symbol being initialized
as there's no inherent reason for that in most cases. In cases where
there's reason for that we should just force the initialization.