SI-107 Adds test for problematic byte to long conversion
While working on another issue I found some questionable looking code
in GenICode#adapt. A bit of code spelunking found that it fixed
issue 107, but that commit didn't include a test. So this commit adds
a test and a comment so the next person doesn't break it.

The code still looks questionable so I've created issue
https://issues.scala-lang.org/browse/SI-7159 to track it.