SI-6900 Fix tailrec for dependent method types.
Uncurry's info transformer could generate a MethodType
with cloned parameter symbols. This type was used
for the LabelDef generated in the TailCalls phase.
But, the RHS of the method still contains types that
refer to the original parmameter symbol. Spurious,
type errors ensued.

This commit changes Uncurry's tree transformer to
keep the value- and type-parameter symbols in the
DefDef consistent with the results of the InfoTransformer.