SI-6900 Fix tailrec for dependent method types
Uncurry's info transformer could generate a MethodType
with cloned parameter symbols. This type was used
for the LabelDef generated in the TailCalls phase.
But, the RHS of the method still contains types that
refer to the original parmameter symbol. Spurious
type errors ensued.

I've spent a good chunk of time pursuing a more
principled fix, in which we keep the symbols in the
tree in sync with those in the MethodType. You can
relive the procession of false dawns:

 https://github.com/scala/scala/pull/2248

Ultimately that scheme was derailed by a mismatch
between the type parameter `T` and the skolem `T&`
in the example below.

    trait Endo[A] { def apply(a: => A): A }
    class Test {
      def foo[T] =
        new Endo[(T, Unit)] {
          def apply(v1: => (T, Unit)) = v1 // no bridge created
        }
    }

Interestingly, by removing the caching in SingleType,
I got past that problem. But I didn't characterize it
further.

This commit sets asides the noble goal of operating in
the world of types, and sledgehammers past the crash by
casting the arguments to and the result of the label jump
generated in TailCalls.